application/requirements.txt
Pillow==8.2.0
opencv_python==4.5.2.52
numpy==1.20.3
scikit-image==0.18.1
imutils==0.5.4
dlib==19.22.0

PyYAML==5.4.1
elasticsearch==7.10.1
requests==2.25.1
uwsgi==2.0.19.1
protobuf==3.17.0
grpcio==1.37.1
pymongo==3.11.4

Flask==2.0.0
Flask_Cors==3.0.10
Flask_RESTful==0.3.9

shapely==1.7.1
Cerberus==1.3.4